refactor BatchData to be in batch.py module
