+++ great_expectations/execution_engine/pandas_batch_data.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from great_expectations.execution_engine.execution_engine import BatchData
+from great_expectations.core.batch import BatchData
 
 
 class PandasBatchData(BatchData):
