+++ great_expectations/execution_engine/sparkdf_batch_data.py
@@ -1,4 +1,4 @@
-from great_expectations.execution_engine.execution_engine import BatchData
+from great_expectations.core.batch import BatchData
 
 
 class SparkDFBatchData(BatchData):
